syntax: Rename deriving/cmp/* to match their current names
